Update: Use normal simple dual ported RAM in the PCIe completion tracker (#1406)
The previous implementation used a 2R x 2W memory for tracking reads
and completions. The updated algorithm dramatically reduces the number
of tracking memory writes needed and shares a single write port for
all updates. Most of the tracker's LUT usage was implementing the
dual write port logic and is now eliminated.

Also, restructure the code that sets the tile/port specific buffer
properties so it is human readable.