Pick a default OFSS top-level file to match the board target (#1436)
When no --ofss argument is set, build_top.sh now searches for an OFSS file to
match the board. Before this change, the build script generated invalid
configurations for most board targets when no --ofss was set. With the
change, any reference board will automatically be configured with IP to
match, simply by setting the board target and no OFSS.